refactor : script 폴더의 위치를 루트 에셋 폴더로 변경
