+++ Assets/Scripts/Actor.cs.meta
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1062eca6e055003f7f1c77035eab8340934a9a783ef486ba5eefc0acdc4f0b57
+size 59
