+++ Assets/Scripts/MapConfig.cs
