+++ Assets/Scripts/StatManager.cs
