+++ Assets/Scripts/WeaponManager.cs
